Documented exceptions thrown by FileDisposeHelper. Restructured test data so that TestDataCopyHelper became obsolete. Restructured tests so that DirectoryDisposeHelper became obsolete. Extended FileUtils with IsValidFilePath.
Related to Issue [WTI-666]


git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@4000 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 7bed29310919ea82c6700affd257a7e68b942633
Former-commit-id: 418f990752da2b746fa44a116f2266b30835f853

+++ Core/Common/test/Core.Common.TestUtil/FileDisposeHelper.cs
@@ -46,6 +46,8 @@ namespace Core.Common.TestUtil
         /// Creates a new instance of <see cref="FileDisposeHelper"/>.
         /// </summary>
         /// <param name="filePaths">Paths of the files that will be created, if the path is valid.</param>
+        /// <exception cref="ArgumentException">Thrown when one of the files in <paramref name="filePaths"/> could 
+        /// not be created by the system.</exception>
         public FileDisposeHelper(IEnumerable<string> filePaths)
         {
             files = filePaths;
@@ -55,14 +57,31 @@ namespace Core.Common.TestUtil
         /// <summary>
         /// Creates a new instance of <see cref="FileDisposeHelper"/>.
         /// </summary>
-        /// <param name="filePath">Path of the single file that will be created, if valid.</param>
+        /// <param name="filePath">Path of the file that will be created, if valid.</param>
+        /// <exception cref="ArgumentException">Thrown when the file could not be created by the system.</exception>
         public FileDisposeHelper(string filePath) : this(new [] { filePath })
         {
+        }
+
+        public void Dispose()
+        {
+            foreach (var file in files)
+            {
+                try
+                {
+                    DeleteFile(file);
+                }
+                catch
+                {
+                    // ignored
+                }
+            }
         }
 
         /// <summary>
         /// Creates the temporary files.
         /// </summary>
+        /// <exception cref="ArgumentException">Thrown when the file could not be created by the system.</exception>
         private void Create()
         {
             foreach (var file in files)
@@ -71,30 +90,38 @@ namespace Core.Common.TestUtil
             }
         }
 
-        public void Dispose()
-        {
-            foreach (var file in files)
-            {
-                DeleteFile(file);
-            }
-        }
-
         /// <summary>
         /// Creates a file at at the given file path. If the <see cref="filePath"/> is
         /// invalid, no file is created.
         /// </summary>
         /// <param name="filePath">Path of the new file.</param>
+        /// <exception cref="ArgumentException">Thrown when either:
+        /// <list type="bullet">
+        /// <item>The caller does not have the required permission.-or- path specified a file that is read-only.</item>
+        /// <item>The specified path, file name, or both exceed the system-defined maximum length. For example, on 
+        /// Windows-based platforms, paths must be less than 248 characters, and file names must be less than 260
+        /// characters.</item>
+        /// <item>The specified path is invalid (for example, it is on an unmapped drive).</item>
+        /// <item> An I/O error occurred while creating the file.</item>
+        /// </list>
+        /// </exception>
         private static void CreateFile(string filePath)
         {
-            try
+            if (FileUtils.IsValidFilePath(filePath))
             {
-                FileUtils.ValidateFilePath(filePath);
+                try 
+                {
+                    using (File.Create(filePath)) {}
+                }
+                catch (Exception e)
+                {
+                    if (e is DirectoryNotFoundException || e is IOException || e is NotSupportedException || e is UnauthorizedAccessException)
+                    {
+                        throw new ArgumentException(e.Message);
+                    }
+                    throw;
+                }
             }
-            catch (ArgumentException)
-            {
-                return;
-            }
-            using (File.Create(filePath)) {}
         }
 
         /// <summary>
@@ -102,17 +129,37 @@ namespace Core.Common.TestUtil
         /// invalid, no file is deleted (obviously).
         /// </summary>
         /// <param name="filePath">Path of the file to delete.</param>
+        /// <exception cref="ArgumentException">Thrown when either:
+        /// <list type="bullet">
+        /// <item>The specified <paramref name="filePath"/> is invalid (for example, it is on an unmapped drive).</item>
+        /// <item>The specified <paramref name="filePath"/> is in use. -or-There is an open handle on the file, 
+        /// and the operating system is Windows XP or earlier. This open handle can result from enumerating 
+        /// directories and files. For more information, see How to: Enumerate Directories and Files.</item>
+        /// <item><paramref name="filePath"/> is in an invalid format.</item>
+        /// <item>The specified <paramref name="filePath"/> exceed the system-defined 
+        /// maximum length. For example, on Windows-based platforms, paths must be less than 248 characters, and 
+        /// file names must be less than 260 characters.</item>
+        /// <item>The caller does not have the required permission.-or- 
+        /// <paramref name="filePath"/> is a directory.-or- <paramref name="filePath"/> specified a read-only file.</item>
+        /// </list>
+        /// </exception>
         private static void DeleteFile(string filePath)
         {
-            try
+            if (FileUtils.IsValidFilePath(filePath))
             {
-                FileUtils.ValidateFilePath(filePath);
+                try
+                {
+                    File.Delete(filePath);
+                }
+                catch (Exception e)
+                {
+                    if (e is DirectoryNotFoundException || e is IOException || e is NotSupportedException || e is UnauthorizedAccessException)
+                    {
+                        throw new ArgumentException(e.Message);
+                    }
+                    throw;
+                }
             }
-            catch (ArgumentException)
-            {
-                return;
-            }
-            File.Delete(filePath);
         }
     }
 }
